vowelsContainer created
